Refactor variable translation; optimise imports.
